refactor: reorganize components folder structure for better organization
- Move input folder from ui/ to components/ root level
- Move Button component from atomic/ to input/ folder
- Create new chat/ folder for chat-related components
- Move AgentPill and ThinkingIndicator to chat/ folder
- Move ThemeToggle to showcase/ folder (non-production component)
- Remove redundant ui/ and atomic/ folders
- Update all import statements and create proper index files
- Fix component showcase import example path
