+++ apps/desktop/src/components/chat/AgentPill.tsx
