+++ apps/desktop/src/components/chat/ThinkingIndicator.tsx
