+++ apps/desktop/src/components/chat/index.ts
@@ -1,0 +1,2 @@
+export { AgentPill } from "./AgentPill";
+export { ThinkingIndicator } from "./ThinkingIndicator";
