+++ apps/desktop/src/components/input/ConversationModeToggleDisplay.tsx
