+++ apps/desktop/src/components/input/MessageInputDisplay.tsx
