+++ apps/desktop/src/components/input/index.ts
@@ -5,6 +5,7 @@
  * These are pure display components without interactive functionality.
  */
 
+export { Button } from "./Button";
 export { ConversationModeToggleDisplay } from "./ConversationModeToggleDisplay";
 export { InputContainerDisplay } from "./InputContainerDisplay";
 export { MessageInputDisplay } from "./MessageInputDisplay";
