+++ apps/desktop/src/components/showcase/index.ts
@@ -1,0 +1,2 @@
+export { ShowcaseLayout } from "./ShowcaseLayout";
+export { ThemeToggle } from "./ThemeToggle";
