--- apps/desktop/src/components/ui/atomic/index.ts
@@ -1,14 +1,0 @@
-/**
- * Atomic UI components barrel export.
- *
- * Re-exports all atomic UI components that are self-contained with no internal
- * state or complex interactions. These are pure display components that render
- * consistently based on props.
- *
- * @module components/ui/atomic
- */
-
-export { AgentPill } from "./AgentPill";
-export { Button } from "./Button";
-export { ThemeToggle } from "./ThemeToggle";
-export { ThinkingIndicator } from "./ThinkingIndicator";
